fix(cost-tracking): keep sub-cent stage costs visible for cheap models
The episode cost breakdown stored and reported money as integer cents, so
any stage cost under half a cent rounded to 0 and rendered as "—". Cheap
models like deepseek-v4-flash never reach a whole cent for scoring or recap,
hiding their costs despite the UI formatting to four decimals.

Add CostEstimator.estimateLlmCostCentsExact returning fractional cents, and
thread Double through the cost breakdown DTOs. buildCosts now recomputes each
LLM stage cost from tokens at full precision, falling back to persisted cents
when the model rate is unknown or there are no tokens. Persistence, aggregate
totals, and the LLM cost gate keep using rounded integer cents; no migration.

+++ src/main/kotlin/com/aisummarypodcast/podcast/PodcastDtos.kt
@@ -127,18 +127,19 @@ data class LlmStageCostResponse(
     val calls: Int,
     val inputTokens: Int,
     val outputTokens: Int,
-    val costCents: Int
+    // Fractional cents so sub-cent stage costs from cheap models stay visible.
+    val costCents: Double
 )
 
 data class TtsCostResponse(
     val model: String?,
     val characters: Int,
-    val costCents: Int
+    val costCents: Double
 )
 
 data class ResearchCostResponse(
     val calls: Int,
-    val costCents: Int
+    val costCents: Double
 )
 
 data class EpisodeCostsResponse(
@@ -148,7 +149,7 @@ data class EpisodeCostsResponse(
     val recap: LlmStageCostResponse,
     val tts: TtsCostResponse,
     val research: ResearchCostResponse,
-    val totalCostCents: Int
+    val totalCostCents: Double
 )
 
 data class UpdateScriptRequest(
